requirements.txt
pyrogram>=2.0.0
TgCrypto>=1.2.0
telethon>=1.28.0
python-dotenv>=1.0.0
requests>=2.28.0
pytz>=2023.3
pillow>=9.0.0
python-socks>=2.3.0
